feat: GachaTradeListenerにBukkitTradeRuleとのインターフェースを貼る
