+++ src/main/scala/com/github/unchama/seichiassist/subsystems/tradesystems/subsystems/gachatrade/System.scala
@@ -4,7 +4,17 @@ import cats.effect.ConcurrentEffect
 import com.github.unchama.seichiassist.meta.subsystem.Subsystem
 import com.github.unchama.seichiassist.subsystems.gacha.GachaAPI
 import com.github.unchama.seichiassist.subsystems.gacha.domain.CanBeSignedAsGachaPrize
+import com.github.unchama.seichiassist.subsystems.gacha.domain.gachaprize.GachaPrize
+import com.github.unchama.seichiassist.subsystems.tradesystems.domain.TradeRule
 import com.github.unchama.seichiassist.subsystems.tradesystems.subsystems.gachatrade.bukkit.listeners.GachaTradeListener
+import com.github.unchama.seichiassist.subsystems.tradesystems.subsystems.gachatrade.bukkit.traderules.{
+  BigOrRegular,
+  BukkitTrade
+}
+import com.github.unchama.seichiassist.subsystems.tradesystems.subsystems.gachatrade.domain.{
+  GachaListProvider,
+  GachaTradeRule
+}
 import org.bukkit.entity.Player
 import org.bukkit.event.Listener
 import org.bukkit.inventory.ItemStack
@@ -16,9 +26,20 @@ object System {
   ): Subsystem[F] = {
     implicit val canBeSignedAsGachaPrize: CanBeSignedAsGachaPrize[ItemStack] =
       gachaAPI.canBeSignedAsGachaPrize
+    implicit val gachaListProvider: GachaListProvider[F, ItemStack] =
+      new GachaListProvider[F, ItemStack] {
+        override def readGachaList: F[Vector[GachaPrize[ItemStack]]] = gachaAPI.list
+      }
+    val gachaTradeRule: GachaTradeRule[ItemStack] = new GachaTradeRule[ItemStack] {
+      override def ruleFor(
+        playerName: String,
+        gachaList: Vector[GachaPrize[ItemStack]]
+      ): TradeRule[ItemStack, BigOrRegular] =
+        new BukkitTrade(playerName, gachaList)
+    }
 
     new Subsystem[F] {
-      override val listeners: Seq[Listener] = Seq(new GachaTradeListener[F])
+      override val listeners: Seq[Listener] = Seq(new GachaTradeListener[F](gachaTradeRule))
     }
   }
